+++ src/auth/auth.service..ts
@@ -4,14 +4,12 @@ import { UsersService } from 'src/users/services/users.service';
 import * as bcrypt from 'bcrypt';
 // import { PinoLogger } from 'nestjs-pino';
 import { LoginDto } from 'src/common/dtos/login.dto';
-import { ConfigService } from '@nestjs/config';
 
 @Injectable()
 export class AuthService {
   constructor(
     private readonly usersService: UsersService,
     private readonly jwtService: JwtService,
-    private readonly configService: ConfigService,
     // private readonly logger: PinoLogger,
   ) {
     // this.logger.setContext(AuthService.name);
