--- src/profile/profile.module.ts
@@ -1,13 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ProfileService } from './profile.service';
-import { ProfileController } from './profile.controller';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { Profile } from './entities/profile.entity';
-import { ProfileRepository } from './profile.repository';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([Profile])],
-  controllers: [ProfileController],
-  providers: [ProfileService, ProfileRepository],
-})
-export class ProfileModule {}
