--- src/profile/profile.repository.ts
@@ -1,15 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { BaseAbstractRepository } from 'src/base/base.abstract.repository';
-import { PROFILE_PAGINATE_CONFIG, Profile } from './entities/profile.entity';
-import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
-
-@Injectable()
-export class ProfileRepository extends BaseAbstractRepository<Profile> {
-  constructor(
-    @InjectRepository(Profile)
-    private readonly profileRepository: Repository<Profile>,
-  ) {
-    super(profileRepository, PROFILE_PAGINATE_CONFIG);
-  }
-}
