--- src/profile/profile.service.spec.ts
@@ -1,19 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ProfileService } from './profile.service';
-import { ProfileRepository } from './profile.repository';
-
-describe('ProfileService', () => {
-  let service: ProfileService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [ProfileService, ProfileRepository],
-    }).compile();
-
-    service = module.get<ProfileService>(ProfileService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
