+++ src/users/dto/user.dto.ts
@@ -1,5 +1,12 @@
 import { ApiProperty, PartialType } from '@nestjs/swagger';
-import { IsEmail, IsNotEmpty, IsString } from 'class-validator';
+import {
+  IsEmail,
+  IsNotEmpty,
+  IsNumber,
+  IsString,
+  MinLength,
+} from 'class-validator';
+import { Level, Position } from '../repository/entities/user.entity';
 // import { Profile } from 'src/profile/entities/profile.entity';
 
 export class UserDto {
@@ -19,21 +26,62 @@ export class UserDto {
   })
   @IsString()
   @IsNotEmpty()
+  @MinLength(6)
   password: string;
 
-  // @ApiProperty({
-  //   description: 'Profile',
-  //   example: {
-  //     username: 'John23',
-  //     name: 'John',
-  //     lastname: 'Doe',
-  //     age: 25,
-  //     telephone_number: '666303030',
-  //     level: '4,00',
-  //   },
-  //   required: true,
-  // })
-  // profile: Profile;
+  @ApiProperty({
+    description: 'Username',
+    example: 'John23',
+    required: true,
+  })
+  @IsString()
+  username: string;
+
+  @ApiProperty({
+    description: 'Name',
+    example: 'John',
+    required: true,
+  })
+  @IsString()
+  name: string;
+
+  @ApiProperty({
+    description: 'Lastname',
+    example: 'Doe',
+    required: true,
+  })
+  @IsString()
+  lastname: string;
+
+  @ApiProperty({
+    description: 'Age',
+    example: 25,
+    required: true,
+  })
+  @IsNumber()
+  age: number;
+
+  @ApiProperty({
+    description: 'Telephone number',
+    example: '666303030',
+    required: true,
+  })
+  @IsString()
+  telephone_number: string;
+
+  @ApiProperty({
+    description: 'Level',
+    example: '4,00',
+    required: false,
+  })
+  level?: Level;
+
+  @ApiProperty({
+    description: 'Position',
+    example: 'BOTH',
+    required: false,
+  })
+  position?: Position;
 }
 
 export class CreateUserDto {
@@ -53,21 +101,62 @@ export class CreateUserDto {
   })
   @IsString()
   @IsNotEmpty()
+  @MinLength(6)
   password: string;
 
-  // @ApiProperty({
-  //   description: 'Profile',
-  //   example: {
-  //     username: 'John23',
-  //     name: 'John',
-  //     lastname: 'Doe',
-  //     age: 25,
-  //     telephone_number: '666303030',
-  //     level: '4,00',
-  //   },
-  //   required: true,
-  // })
-  // profile: Profile;
+  @ApiProperty({
+    description: 'Username',
+    example: 'John23',
+    required: true,
+  })
+  @IsString()
+  username: string;
+
+  @ApiProperty({
+    description: 'Name',
+    example: 'John',
+    required: true,
+  })
+  @IsString()
+  name: string;
+
+  @ApiProperty({
+    description: 'Lastname',
+    example: 'Doe',
+    required: true,
+  })
+  @IsString()
+  lastname: string;
+
+  @ApiProperty({
+    description: 'Age',
+    example: 25,
+    required: true,
+  })
+  @IsNumber()
+  age: number;
+
+  @ApiProperty({
+    description: 'Telephone number',
+    example: '666303030',
+    required: true,
+  })
+  @IsString()
+  telephone_number: string;
+
+  @ApiProperty({
+    description: 'Level',
+    example: '4,00',
+    required: false,
+  })
+  level?: Level;
+
+  @ApiProperty({
+    description: 'Position',
+    example: 'BOTH',
+    required: false,
+  })
+  position?: Position;
 }
 
 export class UpdateUserDto extends PartialType(UserDto) {}
