+++ src/users/repository/entities/user.entity.ts
@@ -2,8 +2,37 @@ import { Expose } from 'class-transformer';
 // import * as bcrypt from 'bcrypt';
 import { PaginateConfig } from 'nestjs-paginate';
 import { BaseEntity } from 'src/base/base.entity';
-import { Profile } from 'src/profile/entities/profile.entity';
-import { Column, Entity, OneToOne } from 'typeorm';
+import { Column, Entity } from 'typeorm';
+
+export enum Position {
+  DRIVE = 'DRIVE',
+  REVES = 'REVES',
+  BOTH = 'BOTH',
+}
+
+export enum Level {
+  '1,00' = '1,00',
+  '1,25' = '1,25',
+  '1,50' = '1,50',
+  '1,75' = '1,75',
+  '2,00' = '2,00',
+  '2,25' = '2,25',
+  '2,50' = '2,50',
+  '2,75' = '2,75',
+  '3,00' = '3,00',
+  '3,25' = '3,25',
+  '3,50' = '3,50',
+  '3,75' = '3,75',
+  '4,00' = '4,00',
+  '4,25' = '4,25',
+  '4,50' = '4,50',
+  '4,75' = '4,75',
+  '5,00' = '5,00',
+  '5,25' = '5,25',
+  '5,50' = '5,50',
+  '5,75' = '5,75',
+  '6,00' = '6,00',
+}
 
 @Entity({ name: 'users' })
 export class User extends BaseEntity {
@@ -14,8 +43,26 @@ export class User extends BaseEntity {
   @Column()
   password: string;
 
-  @OneToOne(() => Profile, (profile: Profile) => profile.user)
-  profile?: Profile;
+  @Column()
+  username: string;
+
+  @Column()
+  name: string;
+
+  @Column()
+  lastname: string;
+
+  @Column()
+  age: number;
+
+  @Column()
+  telephone_number: string;
+
+  @Column()
+  level?: Level;
+
+  @Column()
+  position?: Position;
 
   // async validatePassword(password: string): Promise<boolean> {
   //   return bcrypt.compare(password, this.password);
@@ -23,8 +70,34 @@ export class User extends BaseEntity {
 }
 
 export const USER_PAGINATE_CONFIG: PaginateConfig<User> = {
-  sortableColumns: ['email', 'id', 'createdAt', 'updatedAt', 'deletedAt'],
+  sortableColumns: [
+    'id',
+    'createdAt',
+    'updatedAt',
+    'deletedAt',
+    'email',
+    'username',
+    'name',
+    'lastname',
+    'age',
+    'telephone_number',
+    'level',
+    'position',
+  ],
   nullSort: 'last',
   defaultSortBy: [['id', 'DESC']],
-  searchableColumns: ['id', 'createdAt', 'updatedAt', 'deletedAt'],
+  searchableColumns: [
+    'id',
+    'createdAt',
+    'updatedAt',
+    'deletedAt',
+    'email',
+    'username',
+    'name',
+    'lastname',
+    'age',
+    'telephone_number',
+    'level',
+    'position',
+  ],
 };
